Crate searcher_txt

minigrep lib

A library for my minigrep project

Structs

Config

Contains the fields for the search engine/arguments

Functions

run

Runs the search engine, reads the file specified in the filename argument and outputs the lines of the file that contain the query

search

Searches the query in a file